break out decoration and host cpp files from ui
