+++ CMakeLists.txt
@@ -23,7 +23,10 @@ add_executable(bahnhof
     src/game/main.cpp
     src/game/resources.cpp
     src/ui/ui.cpp
+    src/ui/host.cpp
     src/ui/panel.cpp
+    src/ui/element.cpp
+    src/ui/decoration.cpp
     src/ui/button.cpp
     src/ui/table.cpp
     src/ui/tableline.cpp
