+++ include/bahnhof/ui/tables.h
@@ -160,6 +160,4 @@ private:
     Train& train;
 };
 
-SDL_Rect rendertrainicons(Rendering* r, InterfaceManager& ui, TrainInfo info, SDL_Rect rect);
-
 }
